match the web editor example exactly?
I'm ok with what it was before but this matches the web editor example we have in ml5 more precisely which may reduce confusion? Also, if we removing the square constraint most webcams have a 4:3 aspect ratio -- obviously there are exceptions but this will reduce the number of squished images people see?