Create opencv-keras.md
Added opencv example code so users can use there own web-camera in conjunction with the keras model. Similar to how the website does it with JavaScript. I'm not sure how to add the link to website but if someone shows me. I am more then willing to do the work.